get workers check
